backend/pyproject.toml
[project]
name = "chronic-disease-management"
version = "0.1.0"
description = "Multi-Tenant AI SaaS for chronic disease management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioboto3>=13.0.0",
    "argon2-cffi>=25.1.0",
    "arq>=0.25.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "openai>=2.29.0",
    "passlib[argon2]>=1.7.4",
    "pdfplumber>=0.11.9",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "snowflake-id-toolkit>=0.6.0",
    "sqlalchemy>=2.0.48",
    "tenacity>=9.1.4",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
